Small restructure, fix most compilation errors

+++ Source/Plugins/PhaseCalculator/PhaseCalculator.h
@@ -271,8 +271,8 @@ private:
     Band band;
 
     // filter passband
-    double highCut;
-    double lowCut;
+    float highCut;
+    float lowCut;
 
     // event channel to watch for phases to plot on the canvas (-1 = none)
     int visEventChannel;
